runtak/requirements.txt
Django>=4.2
requests
celery
redis
django-celery-beat
pillow
python-dotenv
